requirements.txt
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
idna==3.4
numpy==1.24.3
pandas==2.0.1
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
python-decouple==3.8
python-gitlab==3.14.0
pytz==2023.3
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.2
